chore: inconsistency cleanup
